IR/requirements.txt
Django~=4.2.6
djangorestframework~=3.14.0
peewee~=3.17.3
pandas~=1.5.3
scikit-learn~=1.4.2
nltk~=3.8.1
autocorrect~=2.6.1
num2words~=0.5.6
numpy~=1.26.4
toolz~=0.12.0
pyspellchecker==0.8.1
spellchecker~=0.4
atpbar~=2.0.0
ir-measures~=0.3.3
python-dotenv~=1.0.1
country-named-entity-recognition~=0.4
beautifulsoup4~=4.11.1
spacy~=3.7.4
numerizer~=0.2.3
tldextract~=5.1.2
pyspark~=3.5.1
matplotlib~=3.9.0
python-dateutil~=2.8.2